+++ kernel/mem/k_phys.c
@@ -5,9 +5,9 @@ extern struct k_mem_plist_t k_mem_physical_pages;
 
 void k_mem_SortFreePages()
 {
-    uint32_t first_free_entry_index = 0;
+    size_t first_free_entry_index = 0;
 
-    for(uint32_t entry_index = k_mem_physical_pages.first_used_page; entry_index <= k_mem_physical_pages.last_used_page; entry_index++)
+    for(size_t entry_index = k_mem_physical_pages.first_used_page; entry_index <= k_mem_physical_pages.last_used_page; entry_index++)
     {
         /* entry in the used pages list that refers to the current page */
         struct k_mem_uppentry_t *used_entry = k_mem_physical_pages.used_pages + entry_index;
@@ -18,7 +18,7 @@ void k_mem_SortFreePages()
             continue;
         }
 
-        uint32_t page_address = entry_index << K_MEM_4KB_ADDRESS_SHIFT;
+        uintptr_t page_address = entry_index << K_MEM_4KB_ADDRESS_SHIFT;
         used_entry->pid_index = first_free_entry_index;
         k_mem_physical_pages.free_pages[first_free_entry_index] = page_address;
         first_free_entry_index++;
@@ -41,9 +41,9 @@ void k_mem_SortFreePages()
     }
 }
 
-uint32_t k_mem_AllocPage(uint32_t flags)
+uintptr_t k_mem_AllocPage(uint32_t flags)
 {
-    uint32_t page_address = K_MEM_INVALID_PAGE;
+    uintptr_t page_address = K_MEM_INVALID_PAGE;
 
     if(k_mem_physical_pages.free_pages_count)
     {
@@ -61,9 +61,9 @@ uint32_t k_mem_AllocPage(uint32_t flags)
     return page_address;
 }
 
-uint32_t k_mem_AllocPages(uint32_t page_count, uint32_t flags)
+uintptr_t k_mem_AllocPages(size_t page_count, uint32_t flags)
 {
-    uint32_t page_address = K_MEM_INVALID_PAGE;
+    uintptr_t page_address = K_MEM_INVALID_PAGE;
     uint32_t entries_sorted = 0;
 
     if(k_mem_physical_pages.free_pages_count && k_mem_physical_pages.free_pages_count >= page_count)
@@ -74,16 +74,16 @@ uint32_t k_mem_AllocPages(uint32_t page_count, uint32_t flags)
         pages we want to allocate. For example, imagine we want to allocate 4 pages and there are only
         6 available pages. There's no point in checking pages after the second page, because there won't 
         be enough pages to service the request anyway  */
-        uint32_t entry_count = k_mem_physical_pages.free_pages_count - page_count;
+        size_t entry_count = k_mem_physical_pages.free_pages_count - page_count;
 
-        for(uint32_t entry_index = 0; entry_index <= entry_count; entry_index++)
+        for(size_t entry_index = 0; entry_index <= entry_count; entry_index++)
         {
-            uint32_t *head_entry = k_mem_physical_pages.free_pages + entry_index;
-            uint32_t first_entry = head_entry[0];
+            uintptr_t *head_entry = k_mem_physical_pages.free_pages + entry_index;
+            uintptr_t first_entry = head_entry[0];
 
             for(uint32_t next_entry_index = 1; next_entry_index < page_count; next_entry_index++)
             {
-                uint32_t second_entry = head_entry[next_entry_index];
+                uintptr_t second_entry = head_entry[next_entry_index];
                 /* likely not the ideal thing to do here. Could be more efficient to try
                 both combinations, and swap the entries if those point to adjacent pages */
                 if(first_entry > second_entry || second_entry - first_entry > K_MEM_4KB_ADDRESS_OFFSET)
@@ -100,23 +100,23 @@ uint32_t k_mem_AllocPages(uint32_t page_count, uint32_t flags)
             {
                 page_address = head_entry[0];
                 struct k_mem_uppentry_t *used_head_entry = k_mem_physical_pages.used_pages + K_MEM_USED_SMALL_PAGE_ENTRY_INDEX(page_address);
-                uint32_t first_page_index = used_head_entry->pid_index;
-                uint32_t next_free_page_index = first_page_index + page_count;
+                size_t first_page_index = used_head_entry->pid_index;
+                size_t next_free_page_index = first_page_index + page_count;
 
                 if(next_free_page_index > k_mem_physical_pages.free_pages_count)
                 {
                     /* error: the allocator somehow "invented" extra pages */
                 }
 
-                uint32_t move_count = k_mem_physical_pages.free_pages_count - next_free_page_index;
-                uint32_t move_offset = next_free_page_index - first_page_index;
+                size_t move_count = k_mem_physical_pages.free_pages_count - next_free_page_index;
+                size_t move_offset = next_free_page_index - first_page_index;
 
                 /* starts at the first page we're allocating */
-                uint32_t *dst_pages = k_mem_physical_pages.free_pages + first_page_index;
+                uintptr_t *dst_pages = k_mem_physical_pages.free_pages + first_page_index;
                 /* starts at the first page after the last page we're allocating */
-                uint32_t *src_pages = k_mem_physical_pages.free_pages + next_free_page_index;
+                uintptr_t *src_pages = k_mem_physical_pages.free_pages + next_free_page_index;
 
-                for(uint32_t move_index = 0; move_index < move_count; move_index++)
+                for(size_t move_index = 0; move_index < move_count; move_index++)
                 {
                     /* we'll move pages from the end of the list to fill the "hole", preserving their order. This is "slow",
                     but allows us to get away with not sorting pages more often */
@@ -143,7 +143,7 @@ uint32_t k_mem_AllocPages(uint32_t page_count, uint32_t flags)
                 used_head_entry[0].pid_index = 0;
                 used_head_entry[0].flags = flags | K_MEM_PAGE_FLAG_HEAD;
 
-                for(uint32_t entry_index = 1; entry_index < page_count; entry_index++)
+                for(size_t entry_index = 1; entry_index < page_count; entry_index++)
                 {
                     /* at first, the kernel process owns those pages. The pid value will be later assigned
                     by the process allocating the page */
@@ -166,7 +166,7 @@ uint32_t k_mem_AllocPages(uint32_t page_count, uint32_t flags)
     return page_address;
 }
 
-uint32_t k_mem_IsValidPage(uint32_t page_address)
+uint32_t k_mem_IsValidPage(uintptr_t page_address)
 {
     if(page_address != K_MEM_INVALID_PAGE)
     {
@@ -176,13 +176,13 @@ uint32_t k_mem_IsValidPage(uint32_t page_address)
     return 0;
 }
 
-void k_mem_FreePages(uint32_t page_address)
+void k_mem_FreePages(uintptr_t page_address)
 {
     page_address &= K_MEM_4KB_ADDRESS_MASK;
 
     if(page_address != K_MEM_INVALID_PAGE)
     {
-        uint32_t entry_index = K_MEM_USED_SMALL_PAGE_ENTRY_INDEX(page_address);
+        size_t entry_index = K_MEM_USED_SMALL_PAGE_ENTRY_INDEX(page_address);
         struct k_mem_uppentry_t *used_entry = k_mem_physical_pages.used_pages + entry_index;
 
         if((used_entry->flags & (K_MEM_PAGE_FLAG_USED | K_MEM_PAGE_FLAG_HEAD)) == (K_MEM_PAGE_FLAG_USED | K_MEM_PAGE_FLAG_HEAD))
